MHW-5 Move socket.io into separate file
